build(facade): refactor JavaExec.commonConfigure to buildSrc
It seems as if I want to import functions, they must in .kt and not .gradle.kts?
